rgw: move the rgw sync code module around
No real code change
