+++ src/rgw/CMakeLists.txt
@@ -51,6 +51,7 @@ set(rgw_a_srcs
   rgw_sync.cc
   rgw_data_sync.cc
   rgw_sync_module.cc
+  rgw_sync_module_log.cc
   rgw_period_history.cc
   rgw_period_puller.cc
   rgw_period_pusher.cc
